move health operation to own file
